(fix) Working boost emails
There's a weird property on Boosts that gets set on export. It uses a counter to loop over the the number of impressions_met

Fixed the template issues.
Made it match the mock
Made the event export the object to set impressions_met
Added a cli function for testing it