(feature) A transcoding endpoint that does something!
Added a new endpoint to v1/media/transcoding
It was relying on the default fall through on /media/ without matching a type. And the front end was just hard coding a response.

This utilizes the AWS S3 client baked into FFMpeg.php to query the configured bucket and list the objects that match the prefix /cinemr_dir/video_guid

Based on the keys that s3 returns, it then builds a TranscodingStatus object which parses the keys into statuses.

A properly transcoded video should have
* guid/source - the raw video
* guid/{video height}.format for each of the presets configured in CONFIG->transcoder->[prefixes]
* thumbnails

If this proves slow or unreliable, the next step is have FFMPEG write status per guid to cassandra